requirements.txt
PyYAML==5.3
SQLAlchemy==1.3.12
beautifulsoup4==4.8.2
metaflow==2.0.1
numpy==1.18.1
pandas==0.25.3
pdfminer==20191125
psycopg2-binary==2.8.4
pylint==2.4.4
pytest==5.3.2
wordninja==2.0.0
tabulate==0.8.6